requirements.txt
Kivy~=2.2.1
requests~=2.31.0
kivymd~=1.1.1
pyzbar~=0.1.9
python-dotenv